+++ changelog.md
@@ -8,6 +8,7 @@ Codion Change Log
 - EntityEditModel.EditorValue.attribute() added.
 - EntityEditor.EditorModels added, related refactoring.
 - EntityEditor.SearchModels added.
+- EntityEditor.replace() added, DefaultEntityEditor.setOrDefaults() now private.
 ### is.codion.swing.framework.model
 - SwingEntityEditModel.createComboBoxModel() memory leak bug fixed, refresh listener now only handles managed column combo box models, related refactoring.
 - SwingEntityEditor added along with SwingEditorModels.